enable more keycloak tests run by IBM partners
(cherry picked from commit 826b171399ac0067115f6c940eb41cef246920b9)
