+++ security/keycloak-oidc-client-extended/src/test/java/io/quarkus/ts/security/keycloak/oidcclient/extended/restclient/TokenPropagationFilterIT.java
@@ -1,5 +1,10 @@
 package io.quarkus.ts.security.keycloak.oidcclient.extended.restclient;
 
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_BASE_PATH;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_FILE;
+import static io.quarkus.ts.security.keycloak.oidcclient.extended.restclient.TokenUtils.USER;
+import static io.quarkus.ts.security.keycloak.oidcclient.extended.restclient.TokenUtils.createToken;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.containsString;
 
@@ -8,17 +13,28 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
+import io.quarkus.test.bootstrap.KeycloakService;
+import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.services.KeycloakContainer;
+import io.quarkus.test.services.QuarkusApplication;
 import io.restassured.response.Response;
 
 @Tag("QUARKUS-3680")
 @QuarkusScenario
-public class TokenPropagationFilterIT extends BaseOidcIT {
+public class TokenPropagationFilterIT {
+
+    @KeycloakContainer(command = { "start-dev", "--import-realm", "--features=token-exchange" })
+    static KeycloakService keycloak = new KeycloakService(DEFAULT_REALM_FILE, DEFAULT_REALM, DEFAULT_REALM_BASE_PATH);
+
+    @QuarkusApplication
+    static RestService app = new RestService()
+            .withProperty("quarkus.oidc.auth-server-url", keycloak::getRealmUrl);
 
     @Test
     public void usernameTest() {
         given()
-                .auth().oauth2(createToken())
+                .auth().oauth2(createToken(keycloak))
                 .when().get("/token-propagation-filter")
                 .then().statusCode(HttpStatus.SC_OK)
                 .body(containsString(USER));
@@ -27,7 +43,7 @@ public class TokenPropagationFilterIT extends BaseOidcIT {
     @Test
     public void jsonUsernameTest() {
         Response response = given()
-                .auth().oauth2(createToken())
+                .auth().oauth2(createToken(keycloak))
                 .when().get("/json-propagation-filter");
         Assertions.assertEquals(HttpStatus.SC_OK, response.statusCode());
         Assertions.assertEquals(USER, response.body().asString());
